+++ CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Corrected  `functioName` field to `functionName` in the `TransitionActionTask` type
+
 ### Added
 - `exh.data.documents.unlinkAllUsers` and `unlinkAllGroups` methods to unlink all users or groups from a document
+- Added `priority` field to the `TransitionActionTask` type
+- Added `TransitionActionTask` to the `AfterActions`
 
 ### Changed
 - RQL `contains` and `excludes` now have their different variations better separated in the type definitions