+++ CHANGELOG.md
@@ -8,10 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Fixed
-- Corrected  `functioName` field to `functionName` in the `TransitionActionTask` type
+- Corrected `functioName` field to `functionName` in the `TransitionActionTask` type
+- `exh.templates.findFirst`, `findById` and `findByName` now correctly state `undefined` can be returned
 
 ### Added
-- `exh.data.documents.unlinkAllUsers` and `unlinkAllGroups` methods to unlink all users or groups from a document
+- Added `exh.data.documents.unlinkAllUsers` and `unlinkAllGroups` methods to unlink all users or groups from a document
 - Added `priority` field to the `TransitionActionTask` type
 - Added `TransitionActionTask` to the `AfterActions`
 - Added `TRANSITION_DOCUMENTS` permission to the `GlobalPermissionName` enum