+++ CHANGELOG.md
@@ -22,6 +22,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - RQL `contains` and `excludes` now have their different variations better separated in the type definitions
 - `exh.data.documents.unlinkUsers` and `unlinkGroups` now also accept an array of user or group ids directly rather than nested in a request body object
   - Thanks to `tran-simon` for the pointing out the initially incorrect `unlinkUsers` type definition!
+- Data service schemas `createMode`, `readMode`, `updateMode` and `deleteMode` accepted values updated
+  - Matching the access mode changes in Data Service 1.4.0
+  - `readMode`, `updateMode` and `deleteMode` now also accept an array of relational modes
 
 ### Deprecated
 - `exh.data.documents.unlinkUsers` usage with an object is deprecated in favor of an array of ids directly or `unlinkAllUsers` for unlinking all users