+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Corrected `functioName` field to `functionName` in the `TransitionActionTask` type
 - `exh.templates.findFirst`, `findById` and `findByName` now correctly state `undefined` can be returned
+- The `notifyAlgoQueueManager` transition action type no longer claims it has a `id` and `version` field
 
 ### Added
 - Added `exh.data.documents.unlinkAllUsers` and `unlinkAllGroups` methods to unlink all users or groups from a document